create/delete vrf file on port events
VRF file is a description of the routing domains served by the
specific agent. This is needed in order for the opflex agent to
understand whether traffic is to be delivered to the external
world and therefore, in certain cases, NATed.

This patch contains the logic that populates these files based on
new Neutron ports joining the managed vSwitch.