Added support for DVS/OVS OpFlex Agent
This adds a derived agent that provides a new parameter
to the plugin, hypervisor_type, which can be used by the
mechanism driver to indicate the appropriate VIF type to
nova when peroforming port binding. It also adds a specialized
Endpoint manager and Firewall driver, in order to support
multiple firewall implementations, and prevent creation of
EP files for ports owned by compute:nova on hosts that are
running nova-compute for vCenter.

Signed-off-by: Thomas Bachman <tbachman@yahoo.com>