Fix RPC invocations to neutron server
There are a few RPCs being invoked on the neutron server as AMQP "cast"
but which should be "call" invocations instead. These RPC invocations
need to reach exactly one neutron server when deployment contains
more than one neutron server (in a multi-node configuration) but ends
up going to every neutron server on account of the current "cast"
invocation with fanout set to True. This leads to the get_gbp_details
method on the neutron server side being processed by every neutron server
and subsequently resulting in undesirable effects such as multiple
host SNAT ports being created instead of exactly one for a given host. This
specific incorrect behavior was actually reported by Cisco IT which led to
the investigation and this patch.