Set hybrid linux bridge ageing to zero
There are reported cases where the linux bridge mac learning breaks datapath
(failed vmotion, where VMs go back and forth 2 hosts is one example).
Since we only have one ingress and one egress port in the LB we set ageing
to zero to avoid this issue altogether. The proper fix for this should
be in Nova, as it's compute agent manages the lifecycle of the
hybrid bridge.

Closes support/issues#591