Support VLAN aware VMs
Add support to Neutron VLAN trunk in opflex networks. For every
port that is part of a Trunk, plug the proper patch ports on the
fabric bridge and write the specified VLAN ID in the EP file.

Also partecipate the VLAN Trunk RPC, by pluggin the newly created
patch ports properly. every new port created this way will be then
processed by the main loop on the subsequent iteration.