Fix metadata host routes
The agent needs to provide the metata host route when optimized
metadata is used. This is needed in cases where the neutron router
isn't connected to the subnet, and therefore ARPs for the default
gateway don't succeed, precluding any attempts to communicate with
the metadata server. In these cases, the DHCP server IP is used
as the next hop (i.e. the same solution used in the reference
implementation for Neutron with isolated networks).