Add support for network updates
The network update RPC is added, along with another RPC that requests
notifications for all ports the agent owns on that network. The network
update notification is used as a trigger for the agent, to let it
know that it needs to query for the ports on the network, as some
non-port parameters may have changed (e.g. DNS, DHCP, etc.), which
get delivered as part of the gbp_details information.