Only handle bound ports
The RPC to get devices for the agent doesn't check to see
if the port is bound to the agent/host. This can lead to the
creation of EP files for ports that the agent isn't responsible
for. A check is added to the device's host. If no host is specified,
or if the host matches the host for the agent, then the port is
processed normally. If there is a host, and it doesn't match the
host for the agent, then the port is removed.

(cherry picked from commit c1c5cd2b17a23ef72556f493842bec0932ab6100)
(cherry picked from commit 54c8ddbf48df59715e080a9641a2776d20d9aa58)
(cherry picked from commit d8b42a4f482302e7545fd88524cb50540d40d273)