request_* RPC API should be of type cast
Set fanout to False to make sure only one server at a time serves
the call.

The previous change was made to prevent all the servers from being
clogged by the agents' requests. However, by using call instead of
cast we also made the RPC blocking, which is not how the agent was
designed. By setting fanout to False, we make sure that only one
server at a time can serve the RPC in an asynchronous and
round robin fashion