Avoid accidental deletion of SNAT EP file
On a fresh start, the agent was relying on the clean-up
of registered endpoints to remove stale SNAT EP files
(i.e. the files that don't have a cooresponding network
namespace). This can cause accidental removal of SNAT
EP file because the files to be removed are calculated
before updates for ports is processed.
Stale SNAT EP files are now tracked separately from
registered endpoints and cleaned up only when they are
not being used.

Signed-off-by: Amit Bose <amitbose@gmail.com>