Support for nested domain details
Provides support for adding lbiface files corresponding
to a Neutron endpoint when the nested domain
VLAN details with are present in the get_gbp_details RPC.

The lbiface files have the following details:
1. uuid (agent assigned, and is unique in the system)
2. interface name
3. VLANs to trunk

E.g.

{
    "uuid": "e4ac02d1-9555-4a37-a09a-236aaf622651",
    "interface-name": "qpicab3d8b8-a3",
    "trunk-vlans": [{"start": 4093}, {"start": 1000, "end": 1001}]
}

Two lbiface files are created for each Neutron endpoint (whenever
the nested domain is configured for the Neutron network),

1. With the interface name corresponding to the Neutron
endpoint (same as which goes into the endpoint file).

2. With the interface name of the uplink interface and the
VLANs detail same as that in (1).

As soon as the agent-ovs sees these lbiface files, its going to
add the necessary flows to trunk the range of VLANs specified
in the "trunk-vlans" section. This should have the effect
of allowing tagged traffic with those VLAN coming from the OpenStack
VMs to go through to the intended destination.

These lbiface files are deleted any time that the corresponding
endpoint file is deleted.