Ensure virtual-ip has CIDR or IPs, not both
The plugin now supports passing allowed-address-pair CIDRs instead
of just /32 host addresses. This resulted in both the CIDR and any
owned addresses in the CIDR to be addeed to the virtual-ip list in
the EP file. While this doesn't break anything, it results in other
agents to log warning messages, as there is no need for both (the
virtual-ip list is used to catch gratuitous ARP packets to the
controller, so it can request exclusive ownership of the IP from
the plugin; since the CIDR already covers the range, any owned /32
IP in the CIDR is redundant). This patch ensures that either the
CIDR or a /32 address will appear in the virtual-ip list, but not
both. Note that CIDRs will never be sent with the "active" flag
set to true in the allowed_address_pairs attribute, so they will
never be added to the ip or anycast_return_ip lists.