Set metadata and dhcp optimization to False for vlan type nets
The service file on opflex agent hosts triggers the creation of flows for the
optimized path. For VM's on vlan type nets we disable the optimized path.
Also added a fix to allow the co-existence a VM on a vlan type net (without
the optimized path) and a VM on a opflex type net (with the optimized path).

(cherry picked from commit 25b62690714db4fb4e3b8952a12b00c408ecf255)